Move data fetching logic to component
